refactor(mcp-io): extract storage and serialization helpers (no behavior change)
